pipelines/scrapping_dest_details/requirements.txt
beautifulsoup4>=4.12.0
requests>=2.30.0
pandas>=2.0.0
google-cloud-bigquery>=3.11.4
google-cloud-storage>=2.9.0
python-dotenv>=1.0.0 